+++ GlobalPayments-iOS-SDK/Classes/Entities/DisputeDocument.swift
@@ -2,8 +2,14 @@ import Foundation
 
 public class DisputeDocument: Encodable, Equatable {
     public let id: String
-    public let type: DocumentType
+    public var type: DocumentType?
+    public var base64Content: Data?
 
+    public init(id: String, content: Data) {
+        self.id = id
+        self.base64Content = content
+    }
+    
     public init(id: String, type: DocumentType) {
         self.id = id
         self.type = type